requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Recommendation systems libraries
implicit>=0.7.0
surprise>=1.1.0
lightfm>=1.16

# Deep learning (optional)
torch>=2.0.0
tensorflow>=2.13.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Interactive demos
streamlit>=1.25.0
gradio>=3.40.0

# Configuration and logging
pyyaml>=6.0
omegaconf>=2.3.0

# Development tools
pytest>=7.4.0
black>=23.0.0
ruff>=0.0.280
pre-commit>=3.3.0

# Progress bars and utilities
tqdm>=4.65.0

# Optional: ML tracking
mlflow>=2.5.0
wandb>=0.15.0

# Optional: Fast similarity search
faiss-cpu>=1.7.4
hnswlib>=0.7.0
